updated bootstrap scripts

+++ scripts/onStart.sh
@@ -5,9 +5,14 @@ set -e
 echo "Starting on Start script"
 
 sudo -i -u ec2-user bash << EOF
-source activate ~/SageMaker/env/rapidsai
-python -m ipykernel install --name 'rapidsai' --user
-echo "Finished setting up rapidsai kernel"
+if [[ -f /home/ec2-user/SageMaker/.create-notebook ]]; then
+    echo "Skipping as currently installing conda env"
+else
+    echo "Installing Jupyter kernel"
+    source activate /home/ec2-user/SageMaker/env/rapidsai
+    python -m ipykernel install --name 'rapidsai' --user
+    echo "Finished setting up Jupyter kernel"
+fi
 EOF
 
 echo "Finishing on Start script"
